feat: Default Libp2p config for node.js and browser peers.
