+++ src/index.js
@@ -41,3 +41,8 @@ export {
   MemoryStorage,
   ComposedStorage
 } from './storage/index.js'
+
+export {
+  DefaultLibp2pOptions,
+  DefaultLibp2pBrowserOptions
+} from './config/libp2p/index.js'
