+++ test/.mocharc.json
@@ -5,5 +5,6 @@
   "exit": true,
   "bail": false,
   "slow": 1000,
-  "exclude": ["test/browser/**/*.js", "test/utils/relay.js"]
+  "exclude": ["test/browser/**/*.js", "test/utils/relay.js"],
+  "timeout": 30000
 }
